#40 Improve on git.properties generation
